refactor source and tooling to make posible add tests ref #65
