+++ .gitignore
@@ -37,6 +37,7 @@
 *.hex
 emulator
 tiny-firmware/skycoin-emulator
+tiny-firmware/test_skycoin-emulator
 **/*.img
 **/*.bin
 !releases/**/*.bin
