+++ Makefile
@@ -87,6 +87,10 @@ emulator: build-deps ## Build emulator
 run-emulator: emulator ## Run wallet emulator
 	./emulator
 
+test: build-deps
+	EMULATOR=1 make -C tiny-firmware/emulator/
+	EMULATOR=1 make -C tiny-firmware/ test
+
 st-flash: ## Deploy (flash) firmware on physical wallet
 	cd tiny-firmware/bootloader/combine/; st-flash write combined.bin 0x08000000
