+++ skycoin-api/check_digest.h
@@ -2,19 +2,9 @@
  * https://github.com/kokke/tiny-regex-c/blob/master/re.c
  * */
 
+#ifndef __SKYCOINAPI_CHECKDIGEST_H__
+#define __SKYCOINAPI_CHECKDIGEST_H__
 
-static bool matchhexdigit(char c)
-{
-  return ((c >= '0') && (c <= '9')) || ((c >= 'a') && (c <= 'f')) || ((c >= 'A') && (c <= 'F'));
-}
+int is_digest(char* digest);
 
-int is_digest(char* digest) {
-    if (strlen(digest) != 64) {
-        return false;
-    }
-    bool bDigest = true;
-    for (int i = 0; i < 64 && bDigest; ++i) {
-        bDigest &= matchhexdigit(digest[i]);
-    }
-    return bDigest;
-}
+#endif  // __SKYCOINAPI_CHECKDIGEST_H__
