+++ tiny-firmware/Makefile.include
@@ -128,16 +128,36 @@ endif
 
 CFLAGS += -DDEBUG_RNG=0
 
+lib$(NAME).a: $(OBJS) $(LIBDEPS)
+	$(AR) rcs lib$(NAME).a $(OBJS) $(LIBDEPS)
 
 ifeq ($(EMULATOR),1)
-$(NAME): $(OBJS) $(LDSCRIPT) $(LIBDEPS)
-	$(LD) -o $(NAME) $(OBJS) $(LDLIBS) $(LDFLAGS)
+$(NAME): $(OBJS) $(LDSCRIPT) $(LIBDEPS) firmware/main.o
+	$(LD) -o $(NAME) firmware/main.o $(OBJS) $(LDLIBS) $(LDFLAGS)
+
+LIBS_DARWIN +=
+TESTLIBS_DARWIN += -L$(CHECK_PATH)/lib -lcheck
+LIBS_LINUX += -L/usr/local/lib/ -lm -lrt
+TESTLIBS_LINUX += -L$(CHECK_PATH)/src -lcheck
+
+ifeq ($(UNAME_S), Linux)
+	LIBS=$(LIBS_LINUX)
+	TESTLIBS=$(TESTLIBS_LINUX)
+endif
+ifeq ($(UNAME_S), Darwin)
+	LIBS=$(LIBS_DARWIN)
+	TESTLIBS=$(TESTLIBS_DARWIN)
+endif
+
+test: proto lib$(NAME).a firmware/test_fsm.o
+	$(LD) -o test_$(NAME) firmware/test_fsm.o $(OBJS) $(LDLIBS) $(LDFLAGS) $(TESTLIBS)
+
 else
 $(NAME).bin: $(NAME).elf
 	$(OBJCOPY) -Obinary $(NAME).elf $(NAME).bin
 
-$(NAME).elf: $(OBJS) $(LDSCRIPT) $(LIBDEPS)
-	$(LD) -o $(NAME).elf $(OBJS) $(LDLIBS) $(LDFLAGS)
+$(NAME).elf: $(LDSCRIPT) lib$(NAME).a firmware/main.o
+	$(LD) -o $(NAME).elf firmware/main.o -l$(NAME) $(LDLIBS) $(LDFLAGS)
 endif
 
 %.o: %.c Makefile
